dev-util/android-studio: add 2024.3.1.14-r1
Revamp the ebuild.
Use QA_FLAGS_IGNORED on the binary files to avoid QA on CFLAGS/LDFLAGS.
Use mv instead doins do avoid duplicating files.
Use studio instead of studio.sh
Move the sysctl.d file to /usr/lib, in line with the other JetBrains packages.
Update the dependencies.

Do not remove plugin binaries. These are meant to be transferred to a target
device. Not run on the host platform. This needs to be fixed on QA side.

Closes: https://bugs.gentoo.org/946601
Closes: https://github.com/gentoo/gentoo/pull/39014
Closes: https://bugs.gentoo.org/941647
Closes: https://bugs.gentoo.org/940458
Closes: https://bugs.gentoo.org/940457
Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>